Duplicate / overlong tweets
Bot was suspended -- was trying to post a duplicate tweet in response
to a user's mention of him. Would hit an error, then fail to register
that the tweet was not new. Line 513 shows the simple fix there.
I also set a hard limit of 10 textual tweets (plus 1 citational tweet)
to the length of any tweet chain. This limit won't be hit often, but it
will stop 30-tweet-long chains from happening. :)
Hopefully Twitter will un-suspend the bot.